// frontend/clients/generated/types/rateModel.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import { Serializer, struct, u16 } from '@metaplex-foundation/umi/serializers';

export type RateModel = {
  kinkUtilBps: number;
  baseBorrowApyBps: number;
  slope1Bps: number;
  slope2Bps: number;
  reserveFactorBps: number;
  maxBorrowApyBps: number;
};

export type RateModelArgs = RateModel;

export function getRateModelSerializer(): Serializer<RateModelArgs, RateModel> {
  return struct<RateModel>(
    [
      ['kinkUtilBps', u16()],
      ['baseBorrowApyBps', u16()],
      ['slope1Bps', u16()],
      ['slope2Bps', u16()],
      ['reserveFactorBps', u16()],
      ['maxBorrowApyBps', u16()],
    ],
    { description: 'RateModel' }
  ) as Serializer<RateModelArgs, RateModel>;
}